requirements.txt
weavel
fastapi-poe
python-dotenv